Revert "Fix Crash in PIN screen."
This reverts commit aaa0df2fb0357788ad5ad8ff2a5a2a05a7fa9e52.

Reason for revert: We are reverting ag/13435976 in a cherry-pick build and this CL depends on ag/13435976 so reverting this as well to avoid a build break. We will not submit this CL (dev is doing a fix forward) but need the revert CL to do the cherry-pick to the .003 build. Will abandon this CL after we have the cherry-pick build.

Bug: 178412096
Change-Id: I4ab2424e70404ab649e714d104de7e5a5a1a4f2f
(cherry picked from commit 6edacfbae62dab344cf9c980f5204047bd9d678a)