Implement QC entry buttons
Quick Control entry points are to replace the StatusBarIcons in the
reference UI. OEM's can config the icons to display by listing their
associated StatusIconControllers in config.xml.

Although StatusBarIcons are removed in AAOS to avoid redundancy, we
should still be prepared for the possibility of re-introducing
StatusBarIcons and letting them coexist. Also, StatusBarIcon
implementation from core Android is to be refactored using the
StatusIcons. Therefore, StatusIconControllers are designed to address
more generic use cases in addition to those required for QC.

Bug: 193032931
Test: Manual + Unit Tests

Change-Id: I2ed4c67741c947e63d2a15954d5b74af4e54f187