DO NOT MERGE Fix failing CarSysUi Tests and flakiness
1. Added car_top_system_bar.xml for simplified top systembar for testing
purposes. We are only testing if the bar is inflated and provided
correctly, and its feature details can be abstracted out for the ease of
testing.

2. Fixed ToastUI test failures from incorrect class instantaition order.

3. Fixed CarSystemBarButtonTests from incorrect method call order and
flakiness from not waiting for thread settling after click event.

4. Made profile picker view optional in TopCarSystemBar -- based on OEM
configuration, TopCarSystemBar might not be used or the profile picker
could be added elsewhere instead.

5. Fixed OverlayViewGlobalStateControllerTest from the addition of the
inset control by the side (in addition to by the type) and process
crashing from test thread flakiness.

Bug:201564704
Bug:201565578
Bug:202055363
Bug:187303085
Test: atest :carsysui-presubmit --include-subdir

Change-Id: I941b09fe92d67f1fc16686329b2b18660afa8a01