Implement temporary fix for PIN button colors being overwritten.
Core Android's KeyguardPinBasedInputView's reloadColors method causes
the color of the PIN buttons (erase and enter) being overwritten
programmatically, causing them to appear deactivated.

For example, overriding the reloadColors method to be a no-op prevents
the colors from being overwritten, but the scope of that change may
introduce some other side-effects.

This CL sets the PIN button colors at the end of the startAnimation
method to the color defined in AAOS, so that the Automotive OEM's have
full control over its configuration, while minimizing the risk of
introducing other side-effects.

Note that this is a temporary fix until we completely decouple
Keyguard's view logic from the core Android implementation.

Bug: 194414715
Test: Manual
Change-Id: I9c723e782623aaf5247906daf7592519af79ccc7