Fix crash in numpad keyguard portrait layout
When inflating the layout for the numpad keyguard on a portrait oriented
device, the inflation fails due to num_pad_key_margin_bottom only
existing in the 'values-land'-folder.
This patch moves the resource to the 'values'-folder so that inflation
works both for landscape and portrait.
Since the resource num_pad_key_margin_horizontal is unused, the entire
dimens file is removed from the values-land folder.

Bug: 202189612
Test:
On a device with portrait display orientation:
atest CtsWindowManagerDeviceTestCases:android.server.wm.KeyguardLockedTests#testLockAndUnlock

Change-Id: Id8337f0966971ab9c3f01840ff86631131c8960e