Add required flag to registerReceiver call in CarVolumeDialogImpl
Android T adds support to allow a runtime receiver to be registered as
not exported, but to ensure apps can take advantage of this, calls to
registerReceiver must specify a flag indicating whether the receiver
should be exported for apps targeting T+ that are registering for
unprotected broadcasts. This commit adds the RECEIVER_EXPORTED
flag to the call to registerReceiver in CarVolumeDialogImpl when
registering for ACTION_CLOSE_SYSTEM_DIALOGS to facilitate testing.

Bug: 161145287
Test: Build
Change-Id: Ia9621ba652468878062803f49c6895303d976ddd