Fix randomly stuck volume dialog
In rare cases the internal state "mShowing" and "mDismissing" are not
properly updated when the animation ends unexpectedly. Then the volume
dialog will not be removed from the screen.
Instead of keeping track in the CarVolumeDialogImpl class the
Dialog.isShowing() attribute is used. mDismissing will be reset when the
dialog is shown again.

Test: Manual tests which throws an exception in animationEnd()
Bug: b/206665825
Change-Id: I02481434d7038d43f2fe35bbbd4c7258071bb38f