Remove empty modules
Remove empty WM and SysUI modules because they are unnecessary. The
GlobalRootComponent already has builders for the WM and SysUI
subcomponents.

While the @Modules.subcomponent annotation can be used to optimize the
graph, in this case it's not needed because these subcomponents are
exposed externally. They are also essential to System UI and cannot be
made optional anyway.

Bug: 205725937
Test: manual
Change-Id: I746b7721333a81078a7614fc20f97525bdae7333