Pass non-IME controls to InsetsController
The InsetSourceControls need to be passed to the InsetsController.
Without the controls, the InsetsController will be unable to properly
animate in/out the system bars. However, because the parent
DisplayImeController handles the ime animations itself, those controls
should not be passed to the InsetsController to prevent access after the
leash is released.

Bug: 229256359
Test: manual, atest DisplaySystemBarsControllerTest
Change-Id: Ifde6e0b8d67536724e8053b8dd42a40d648f9a60