Fix stuck notification shade when opening gesture ends in navigation bar
The Notification shade expand state is not correctly set, when
an opening swipe ends in the navigation bar. In that case the
notification shade is fully expanded and stays so, even when the
finger is lifted. As the system does not track the correct state,
this can lead to a broken UI experience, for example, the shade
is not automatically close if a other component requests focus.

Bug: 203791752
Test: manual testing emulator, Seahawk
Change-Id: Iaed2064fc3a875945a733da9b9f124ee5044d0e5