Complete panel animation only on release
The panel animation should not be automatically completed while the user
is still touching the screen. If the user is still mid-swipe, a jittery
behavior will occurr where the panel will animate up or down to the
completion point and then back to the current touch point on next move.

Bug: 237455064
Test: manual
Change-Id: Iea91a2282b5bb617763523a7927405fa78972d01