Allow overriding of ButtonSelectionStateController
Allow extensions of CarSystemUI to override the default
ButtonSelectionStateController implementation.

Its possible the UI is using TaskViews or DisplayArea or a combination of both. OEM would have custom buttons on the nav bar as per their design and hence finding the top activity will depend on use case specific.
For example, if the Display Area are in WINDOWING_MODE_MULTI_WINDOW the current logic will find the root activity but thats not true for all.

ButtonSelectionStateController logic is still the same. Its just refactored in a way to allow override

Bug: 228639659
Fix: 241463042
Test: Manual
Change-Id: I78badc17250280376dac20e9433e8beaada11a49