Drive Mode Button for the top status bar.
A Quick Control button in the top status bar that shows available drive modes and allows to change them.
As the feature is just for internal testing at the moment and is not connected to any hardware (e.g. no VHAL connection) Drive modes for now are hardcoded and provided by the SingletonDriveModeManager and changing a Drive Mode means applying an RRO that changes the looks of the sytem (e.g. accent color)

Bug: 196526394
Test: Manual
Change-Id: Ie39a59a46c34e387198bd2e11ab7cd1985dbdd08