Make input lock icon work correctly with show/hide pair
Currently if input lock is started or user touches the screen of
a display with input lock enabled, the lock icon is shown and it
will fade out automatically.
If SystemUI starts (after SystemUI crashes or device reboots) while
the display input lock is activated, the lock icon transition
animation may start before the lock icon is actually drawn on the
screen. In that case, the transition animation is not played and
the state of the lock icon is not cleaned up normally, so it is not
shown again when user touches the screen. This CL fixes this to only
fade out if the lock icon has been laid out.
Also ensure input lock icon works correctly when multiple calls are
made to show and hide (fixes multiple touches on the screen of
a display with input lock enabled).

Bug: 253284696
Test: m CarSystemUI
Test: atest CtsCarTestCases:CarDisplayInputLockTest
Test: atest DisplayInputSinkControllerTest

Change-Id: I1f965798f8e06835178ed72926bf26e74c3f5204
Signed-off-by: Dohyun Lee <dohyun.lee@lge.com>