Implement user invisible waiter
This is used to wait until previous user is in invisible state.
When changing user, previous user is stopped, and new user is
started. But new user can not be started if occupant zone is not
unassigned for previous user yet, and occupant zone unassignment is
processed on user invisible event. In this reason, we should wait
until previous user is in invisible state for stable user starting.

Bug: 275973135
Test: m
Test: manual test
1. test path
    1) User A logs in on any passenger display.
    2) Execute user picker on that display.
    3) User B logs in.
2. Expected result
    User B is started and B's Home is launched.

Change-Id: Id03be3e713e1e314e69d451dcec18b5a8dad3a0e