Resolve unresponsive HVAC systemui item
Add exception handling for getProperty. For hvac properties that depend
on HVAC_POWER_ON check if power is on before get/set. Initialize
UI logic to be powered off on boot in case hvac is powered off and
values cannot be retrieved. Add default values for UI elements in case
power is off on boot.

Bug: 276275880
Test: atest HvacControllerTest
Test: atest HvacBooleanToggleButtonTest
Change-Id: I1812bd233727ff12848e20ef81e2afa6934cad14