Fix NPE in QuickControlsStatusIconListController
When QuickControlsStatusIconListController#updateStatus is called,
icon drawable of the sub-controller may be null because it has not been
initialized.
To prevent an NPE, the operation for the sub-controller with a null icon should be skipped.
If the icon of the sub-controller is updated at a leter time,
the operation will be retried.

Bug: 280531399
Test: manual - reboot device and check if SystemUI does not crash.
Change-Id: I688ef13ba99f237d22abcb3d192a5f73534aacd6