Update the UserPicker header state on any lifecycle event
It's possible to get in an incorrect userpicker state for more than just
stopping. For example, if a user is started from outside the userpicker,
it may get stuck in the logged out state. By listening to all the user
lifecycle events and checking the state to see if it needs changing, it
can always stay up-to-date.

Bug: 281769348
Test: manual (switched via settings and userpicker multiple times and verified correct userpicker state after)
Change-Id: Ib30ece817f9a4cbeb883496782e925cb656b3da1