Shift handle bar for top-down panels
When setting the y translation for the handle bar for top-down panels,
it should be shifted up by the height of the view such that the bottom
is aligned with the bottom of the window. Otherwise the top of the view
will start at the bottom of the window and will end up outside the space
in which it is able to render.

Bug: 272515947
Test: manual
Change-Id: I53f7c5c1851baa877c6c155fb152b6cbf0e60554