Fix failing tests after notification direction reverted
In ag/23219195, the CarSystemUITests res config.xml was changed to disabling the top bar, since notifications were changed to open from the bottom in ag/22996088. Now that the notification panel has been reverted back to opening from the top bar ( ag/24679968 ), the test changes should also be reverted. To decrease ambiguity, the config for the enabled status of the bottom bar and the config for which NotificationPanelViewMediator is used are explictly defined in the test resources config.xml file.

Bug: 300720451
Test: atest
CarSystemUITests:com.android.systemui.car.systembar.CarSystemBarControllerTest

Change-Id: I302008fde77b34801edcffc0eb9c5c9d72ccb342