Fix failed ProfileSwitcherTest
Due to a recent change, switchUser() is called in a background thread.
In the test, although the "background" thread is the test thread,
switchUser() is called after the assertion, so the assertion will
fail. To ensure the assertion is called after switchUser(), this CL
posts a Runnable and does the assertion in the Runnable.

Fixes: 300712568
Bug: 299994119
Test: atest com.android.systemui.car.qc.ProfileSwitcherTest
Change-Id: I740f28211d1fb189a824b669ecff0f0cbae00672