Fix leak in NotificationPanelViewController
The NotificationClickHandlerFactory would store up a new
Notification click listener every time the notification view was
reinflated due to a configuration change.
Also changed the CarServiceOnConnectedListener to a member to
reinitialize the setup of the NotificationViewController for every
config change, since otherwise the mNotificationView passed to it would
be stale until the car service would reinitialize itself.

Bug: 301492797
Test: Manuel (memory dump)
Change-Id: I1f5946f3bdc89370586678295c0bc0536dfc8cf9