Update action buttons of system panels triggered from status bar
Make sure that action buttons of "Bluetooth", "Network", "Display
Brightness" and "Profile" panels, which are triggered by corresponding
status bar buttons always use current system language.

IMPORTANT: while not directly related to the issue, it seems that the
`CarSystemBar` should be `@SysUISingleton`. At the moment it is not and
it is instantiated twice. This causes some updates of the system UI
elements triggered from this class to get applied two times. It doesn't
look like intended behavior, but there might be some specific reasons
for it to be the way it is. The change adds the presumably missing
annotation and all seems to work as expected also automated tests keep
passing, this part can be reverted if there are any major concerns with
the change, but it is very likely causing extra visual artifacts and/or
slight delays in some relevant scenarios (profile switch, day/night
changes, etc).

Fix: 319133238
Test: atest CarSystemBarTest, manually - make sure "Profiles and
accounts settings" button in "Profiles" system bar menu always uses
current system language when it is changed via Settings

Change-Id: I6f2982224f493e91a5a887dfcfbe5aa7e21246e6