Overlay keyguard_pin_view directly
In UDC, core uses a wrapping view in keyguard_pin_view to include
portrait and landscape versions. However, this breaks maxWidth and
maxHeight attributes becasue the pin view is not the direct child of
KeyguardSecurityViewFlipper. Overlaying the pin view directly solves
this in UDC while a larger refactor already solved this for main.

Bug: 311297883
Test: manual
Change-Id: Iad2f26782ce2d900a14d7b7a64b3ebdf3c5c5a9b