Exclude Assistant widget test
With the addition of the media_card_fullscreen flag in Launcher, the
Assistant card on Home screen will no longer be visible. The
"testAssistantWidget" test here:
https://source.corp.google.com/h/googleplex-android/platform/superproject/main/+/main:platform_testing/tests/automotive/functional/home/src/android/platform/tests/HomeTest.java;l=44
will fail once the flag is enabled. Since the test is in CatBox and
flagging infrastructure is not supported there yet, the workaround for
now is to disable this test.

Bug: 331719655
Test: atest --test-mapping
packages/apps/Car/SystemUI:auto-end-to-end-presubmit and make sure
testAssistantWidget test is not run.

Change-Id: Id0e0808567d45560faf522bc1d4a36b436cac923