Fix reporting seen notifications
Currently the callback OnUnseenCountUpdateListener is called whenever
panel opens or when a new notification is posted. in the callback
seen notifications are reported regardless if the panel is open which is
not expected. Seen notifications should ony be reported when there is
enough confidance that a notification is seen, which ususlly happens
when the notification center is opened.

This fixes this assumption in
NotificationAssistantServiceTest#testOnNotificationVisibilityChanged

Bug: 282895844
Test: atest CtsLegacyNotification29TestCases:android.app.notification.legacy29.cts.NotificationAssistantServiceTest#testOnNotificationVisibilityChanged --iterations=10
Change-Id: Ia555ed43417da8763787fc9c3a58a7a7979b738b