Wait to show ABA blocking content
The ABA sometimes appears before it launches another blocking activity
depending on the speed of checking for media sessions/in call services.
This change waits to confirm that there are no active media sessions or
calls before displaying the ABA content. Also removed the splash screen
from the ABA for a smoother transition when it comes up.

Bug: 347746984
Flag: EXEMPT minor bug fix
Test: Manual
Change-Id: I260f46ea485c7f0a8da85a729340c8c519ea1de6