Fix scroll bar flickering on user picker for secondary user
Ignore the systemBars insets for UserPickerActivity if the bar is
dismissable.

Add sample overlay to enable bar policy.

The scroll bar flickers when the recyclerView resized on nav bar hide,
in this scenario UserPickerActivity should igonre the insets of nav bar
to avoid resizing.

If the nav bar is configured to be persistent, UserPickerActivity should
take care of the nav bar insets.

Fix: 349475156
Flag: EXEMPT bugfix
Change-Id: I0f31f01fa980742545cfcdb0e386a412a9f5cc19
Test: com.android.systemui.car.userpicker.UserPickerBottomBarTest