Update lock status for visible background users
Basically, the lock status is handled by the keyguard. However, the
keyguard for secondary users in MUMD configuration is not currently
supported. As a result, the device lock status for the visible bg user
is not updated when the user is created. To fix this, it's necessary to
trigger the update of the visible bg user's lock status from somewhere
other than the keyguard.
There are some APIs that can update the lock status, as follows:
- TrustManager#reportEnabledTrustAgentsChanged(int)
- TrustManager#unlockedByBiometricForUser(int, BiometricSourceType)
- TrustManager#clearAllBiometricRecognized(BiometricSourceType, int)

We can choose one of the above APIs to perform an update on the lock
status for the visible background user. As a temporary solution, we
intend to resolve the issue by using #reportEnabledTrustAgentsChanged.

Bug: 348298444
Test: atest --user-type secondary_user_on_secondary_display CtsWidgetTestCases:TextViewReceiveContentTest
      atest --user-type secondary_user_on_secondary_display CtsWidgetTestCases:TextViewTest
(cherry picked from https://partner-android-review.googlesource.com/q/commit:45da6c432f33802e5a3637c85b29b7e9919422c0)
Flag: NONE bugfix
Merged-In: I8eb407095688ea2c031517299d7933dd4a73bc8a
Change-Id: I8eb407095688ea2c031517299d7933dd4a73bc8a