Reorder embedded tasks to top when the task view is coming to the top
Side effect from ag/27029279 in which the call to bring the embedded
tasks to the top was removed. When home is coming to the top due to a
back button press (basically onResume for the task view), the call does
not go through handleRequest in which all the embedded tasks are brought
to the top. Instead, handle this behavior in startAnimation since that
is called after the task has been resolved.

Flag: EXEMPT (bug fix)
Bug: 338639777
Test: Open apps grid, use rotary controller to press back, verify if
maps opened.
Test: atest MultiDisplayClientTests

Change-Id: I00a40520feb8f2f6e2d8f4e17c0db55f210c661c