Add keyguard transitions to car.
KeyguardTransitionRepository was previously not being started at all since it wasn't in the CarKeyguardModule. This caused test failures when keyguard_wm_state_refactor was enabled, since the repository remained in OFF permanently and the refactor depends on transitions to set WM state.

This doesn't appear to have any side effects in manual testing, and passes presubmits.

Bug: 278086361
Test: manual on car emulator
Test: presubmits
Flag: N/A bugfix for flag
Change-Id: I59af56af00c289975a5806c0a68f6d62ba940635