Add wmshell_defaults to CarSystemUI
This will make sure that car SystemUI as a wmshell client app satisfies all the dependencies of WMShell. In particular for protologging, to ensure it includes the viewer config files to decode the log messages.

This will also allow us to more easily make updates to anything the WM shell clients modules need to set up in the build rules.

Bug: 376039977
Flag: EXEMPT build rule update
Change-Id: I1ab4b17b3b19bb4cccff63e37391784f0178494c