Refresh display input lock in onDisplayAdded
There was a timing issue when the display input lock settings value was
parsed before all passenger displays are ready, not recognizing the
unique display id and failing to apply the display input lock.
The issue was observed especially when rebooting the system.
Fixed the issue by re-parsing the dispaly input lock settings in
onDisplayAdded callback.

Bug: 378543659
Flag: EXEMPT bugfix
Test: Manual testing using the reference hardware.
Test: atest DisplayInputSinkControllerTest

Change-Id: I7584300f6d1c74dcab1ecc4fac7ce8afa941ada6