Add force immersive config for SUW
For configurations that enable the use of remote inset controllers to
control system bar behaviors, it may be desired to supplement an
existing policy with the option to force hide one or both system bars
while Setup Wizard is in progress. This change adds a config to provide
this option.

Bug: 409074387
Bug: 437156250
Test: manual
Flag: NONE bugfix
Change-Id: I48637fb0a913325776a284dfdc9537043b2dc7da