Fix ABA config change crash
The BlockerViewModel depends on the ABA onCreate() to initialize its
data, such as the blocked activity component name, etc. If the ABA gets
recreated from a config change, the livedata tries to register the same
source twice which throws an exception. Because BlockerViewModel is tied
to the ABA as its lifecycle owner, we can assume that the blocked
activity remains the same.

Bug: 401607066
Test: manual
Flag: EXEMPT bug_fix
Change-Id: I86b22328ce735af5037f1e00bec1b9acfb9f8df6