Remove stub top bar from userswitcher
The majority of this bar is non-functional in this usage. Removing the
bar and replacing with the only functional component (the clock) will be
clearer to users and be consistent with the MUMD userpicker
functionality.

Bug: 406083485
Test: manual
Flag: NONE bugfix
Change-Id: I017b5515d47f71392c5824795b55f94d760f0dd7
(cherry picked from commit 75786ba16e1fcf1432650a1b1a7bf8c0cfcdfcd9)