Fix: Volume dialog seekbar is not responsive.
When a user is dragging the volume seekbar, the dialog should not
refresh. This is to prevent a loop of events that causes UI
unresponsiveness.

A new flag mIsDragging is introduced to track the dragging state of the
seekbar. The dialog refresh is skipped when this flag is true.

Flag: EXEMPT bug fix
Bug: 353615946
Test: Manual test
Change-Id: I40cd92b83249d1a42e6825d97428318a5383991e