Add gravity support to HUNs and remove scrim logic
Refactors HUN panels to support a `gravity` property for positioning
at the top or bottom of the screen.

This is achieved by adding the `gravity` property to `BasePanel` and
updating `HunWindow` to use it for its WindowManager LayoutParams.
`PanelConfigReader` is updated to create `BasePanel` instances for HUNs.

Also, the obsolete `scrim` logic is removed from the panel update system,
simplifying the implementation.

Bug: 439938419
Test: atest CarSystemUITests
Flag: com.android.car.scalableui.enable_ext_panel_updates
Change-Id: I9f10e84af601d87da473e2636ce856a68b732aed