feat(SystemUI): Dispatch show/hide events in SystemUiWindowBase
This change introduces the ability for `SystemUiWindowBase` to dispatch
show and hide events. This is accomplished by adding `show()` and `hide()`
methods to the `SystemUiWindow` interface and implementing them in
`SystemUiWindowBase`.

The `CarHeadsUpNotificationSystemContainer` is updated to use these new
methods to control the visibility of the HUN window. This allows for a
cleaner separation of concerns and removes the need for the container to
directly manipulate the visibility of the HUN window's root view.

Bug: 447230310
Test: atest CarSystemUITests
Flag: com.android.car.scalableui.enable_ext_panel_updates

Change-Id: I22346251cbeadde67cf5f7c2000b16de8c347fbe