Defer System Bar creation until panel configs are ready
Previously, the System Bar initialization process could race against the
reading of panel configuration files. This could lead to errors or
incorrect layouts if the bars were created before their configuration
was fully loaded.

This change introduces a PanelConfigReadStateMonitor to solve this
race condition. The CarSystemBarControllerImpl now waits for this monitor
to signal that the panel configurations are ready to restart the
creation of the system bars.

The SystemUiConfigurationProvider and SystemUiWindowProvider have
also been updated to be aware of this ready state, preventing them from
providing stale or incomplete data.

Fixes: 455586977
Test: atest CarSystemUITests
Flag: com.android.car.scalableui.enable_ext_panel_updates
Change-Id: I182fe1f7932d8895dda6d1f496ed3fe4dfa10176