Listen to animation cancellation from SurfaceAnimator
When a transition get cancelled (for merge purposes or otherwise), the
SurfaceAnimator update listener will never reach 1.0, as it will stop
in-place at its current progress. The animator should listen for this
cancel signal and perform updates there as well to finalize the panels
into the correct states when this happens.

Bug: 452655192
Test: manual
Flag: NONE bugfix
Change-Id: I1008da8041ab71fa6a801c0d0cda6e852b712a61