Trigger panel empty event when task is moved to back
When a task within a TaskPanel is moved to back, it will be re-parented
out of the panel and into the default TDA. Because this re-parenting
happens before startAnimation, the changes list will not include this
task as a child of the panel RootTask. Therefore, checking the previous
parent token against the RootTask token for this transition is necessary
to detect this situation and send the event accordingly.

Bug: 452305601
Test: manual
Flag: NONE bugfix
Change-Id: I3e1adf5f2b3b9cd2a23c0c187919eb2b6354da0c