Run AutoLayoutManager Tx on WMShell-Main
Updating the AutoLayoutManager Tx on the same thread that process the
SurfaceTx can cause unwanted delay in drawing the frames. They
might be moved to the WmShell Main thread as they are not part
of the same SurfaceTx, hence freeing the SysUI MainThread from further
processing the InputEvents.

Bug: 446233885
Test: Manual
Flag: EXEMPT BUGFIX

Change-Id: Iabb4adcaa45f64dd9988af2b24c50bc438c86572