Change inset visibility later for SUW
If the system is booted while SUW is considered in progress, the system
will attempt to hide these insets very early in the boot sequence,
potentially even before the bar views are available (since WM inits
first). This seemingly creates a race condition where the insets are
thought to be non-visible but they still are.

Since the SUW cannot be considered in-progress until the user is
unlocking anyways, delay the SUW specific hiding until this time.

Bug: 457805952
Test: manual
Flag: NONE bugfix
Change-Id: I617ce049064b07072e9254c773f1e45a2e3ae79b