Force current panel state for conflicts when needed
Under certain circumstances, a conflict resolution event may not resolve
if either the event is not expected to be handled or if the panel is
already in the state it's expected to be in. For these cases, include
the current panel state in the transaction to force the WM to react
appropriately.

Bug: 452380827
Test: manual
Flag: NONE bugfix
Change-Id: Ia52ed2ac6bd20931fbe2016ad47ac3cbf863d48c