Add display density button to v2 toolbar
This change adds a new button to the v2 compatibility toolbar that allows users to adjust the display density of the application.
The new button launches a dialog in the Car Settings app where the user can select a new display density.

Bug: 443760606
Test: manual
Flag: com.android.systemui.car.display_compatibility_v2
Change-Id: Ieaa81e5bea1bb7ca8749c367bba1da0259634f61