Use PanelState for focus selection
PanelState is updated earlier to the upcoming variant than the actual
state of the panel (since the panel may be changed/animated over time).
To prevent timing issues where a panel is focused for being the top
visible layer when this is no longer the case, check the PanelState
instead before setting focus.

Bug: 460836854
Test: manual
Flag: NONE bugfix
Change-Id: I23e35e3452479cdcf22d594445b94641c33f6c4a