Hide TaskPanel when last task is moved to back
Implements the `moveRootTaskToBack` callback in `TaskPanel`, which was
added to the `RootTaskStackListener`.

When `AutoTaskStackControllerImpl` determines that a RootTask managed
by this TaskPanel should be moved to the back (because its last opaque
task was sent to the back), it calls `moveRootTaskToBack`.

This implementation reacts by calling `reportTaskPanelEmpty()`,
triggering the ScalableUI logic to hide or collapse the now-empty panel.

Bug: 409394537
Test: manual
Flag: EXEMPT BUGFIX
Change-Id: Ie9dda9a66f3662a94cfdb830883aad458d0b0581